rename 'utilities' folder to 'data'
